// src/character/enums/character-archetype.enum.ts
/**
 * Архетипы персонажей
 * Основаны на классических архетипах Юнга и дополнительных литературных архетипах
 */
export enum CharacterArchetype {
  // Классические архетипы
  HERO = 'HERO',                 // Герой - смелый, решительный, принципиальный
  MENTOR = 'MENTOR',             // Наставник - мудрый, опытный, помогающий
  CAREGIVER = 'CAREGIVER',       // Заботливый - поддерживающий, оберегающий, внимательный
  REBEL = 'REBEL',               // Бунтарь - своевольный, вызывающий, нонконформист
  LOVER = 'LOVER',               // Любовник - страстный, эмоциональный, чувственный
  EXPLORER = 'EXPLORER',         // Искатель - любопытный, независимый, авантюрный
  CREATOR = 'CREATOR',           // Творец - инновационный, изобретательный, художественный
  RULER = 'RULER',               // Правитель - контролирующий, организованный, лидерский
  JESTER = 'JESTER',             // Шут - веселый, спонтанный, забавный
  SAGE = 'SAGE',                 // Мудрец - аналитический, знающий, глубокий
  INNOCENT = 'INNOCENT',         // Невинный - оптимистичный, чистый, доверчивый
  MAGICIAN = 'MAGICIAN',         // Волшебник - трансформирующий, проницательный, вдохновляющий

  // Дополнительные архетипы
  ORPHAN = 'ORPHAN',             // Сирота - уязвимый, ищущий принадлежности
  WARRIOR = 'WARRIOR',           // Воин - отважный, стойкий, боевой
  SEDUCTRESS = 'SEDUCTRESS',     // Соблазнительница - манипулятивная, привлекательная
  TRICKSTER = 'TRICKSTER',       // Трикстер - хитрый, обманчивый, неоднозначный
  SHADOW = 'SHADOW',             // Тень - темный, подавленный, загадочный
  GUARDIAN = 'GUARDIAN',         // Хранитель - защищающий, верный, преданный
  ARTIST = 'ARTIST',             // Художник - выразительный, эмоциональный, чувствительный
  INTELLECTUAL = 'INTELLECTUAL', // Интеллектуал - рациональный, любопытный, аналитический
  COMPANION = 'COMPANION',       // Компаньон - верный, дружелюбный, поддерживающий
  FEMME_FATALE = 'FEMME_FATALE', // Роковая женщина - опасная, соблазнительная, таинственная
  ANTIHERO = 'ANTIHERO'          // Антигерой - сложный, противоречивый, неоднозначный
}